[POAE7-858] disable RDD cache related PMem intialization as default and add PMem related logic in SparkEnv (#18)
* disable RDD cache related PMem intialization as default

* add PMem related logic in SparkEnv

* fix bug happened when spill fallback to disk

+++ src/main/java/org/apache/spark/util/collection/unsafe/sort/PMemReaderForUnsafeExternalSorter.java
@@ -35,7 +35,7 @@ public final class PMemReaderForUnsafeExternalSorter extends UnsafeSorterIterato
         this.numRecordsRemaining = numRecords - position/2;
         this.taskMetrics = taskMetrics;
         int readBufferSize = SparkEnv.get() == null? 8 * 1024 * 1024 :
-                (int) SparkEnv.get().conf().get(package$.MODULE$.MEMORY_SPILL_PMEM_READ_BUFFERSIZE());
+                (int) (long) SparkEnv.get().conf().get(package$.MODULE$.MEMORY_SPILL_PMEM_READ_BUFFERSIZE());
         logger.info("PMem read buffer size is:" + Utils.bytesToString(readBufferSize));
         this.byteBuffer = ByteBuffer.wrap(new byte[readBufferSize]);
         byteBuffer.flip();
